15:42 - 16/06/2024

+++ models/intermediate/int_goles_jugador.sql
@@ -17,7 +17,7 @@ final as (
     select 
         j.id_jugador,
         j.nombre,
-        count(tipo_evento) as numero_goles,
+        count(tipo_evento) as numero_goles::INT,
         avg(minuto) as minuto
         
     from jugador j
